+++ src/main/scala/xiangshan/frontend/Frontend.scala
@@ -22,6 +22,7 @@ import freechips.rocketchip.diplomacy.{LazyModule, LazyModuleImp}
 import xiangshan._
 import xiangshan.cache._
 import xiangshan.cache.prefetch.L1plusPrefetcher
+import xiangshan.cache.mmu.{TlbRequestIO, TlbPtwIO}
 import xiangshan.backend.fu.HasExceptionNO
 import system.L1CacheErrorInfo
 
@@ -35,7 +36,7 @@ class Frontend()(implicit p: Parameters) extends LazyModule with HasXSParameter{
 
 
 class FrontendImp (outer: Frontend) extends LazyModuleImp(outer)
-  with HasL1plusCacheParameters 
+  with HasL1plusCacheParameters
   with HasXSParameter
   with HasExceptionNO
 {
