+++ src/main/scala/xiangshan/mem/lsqueue/StoreQueueData.scala
@@ -21,7 +21,7 @@ import chisel3.util._
 import utils._
 import xiangshan._
 import xiangshan.cache._
-import xiangshan.cache.{DCacheWordIO, DCacheLineIO, TlbRequestIO, MemoryOpConstants}
+import xiangshan.cache.{DCacheWordIO, DCacheLineIO, MemoryOpConstants}
 import xiangshan.mem._
 import xiangshan.backend.roq.RoqPtr
 
@@ -52,7 +52,7 @@ class SQPaddrModule(numEntries: Int, numRead: Int, numWrite: Int, numForward: In
       data(io.waddr(i)) := io.wdata(i)
     }
   }
-  
+
   // content addressed match
   for (i <- 0 until numForward) {
     for (j <- 0 until numEntries) {
